+++ ZPI_game/Assets/Maps/Map.cs
@@ -44,9 +44,9 @@ namespace Maps
     }
     public struct Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public Point(int x, int y)
+        public float X { get; set; }
+        public float Y { get; set; }
+        public Point(float x, float y)
         {
             X = x;
             Y = y;
